Clean up score contribution calculations and display results as a table
